editora feito e livro quase feito

+++ src/br/livraria/dao/PesquisaDao.java
@@ -6,12 +6,14 @@ import java.util.List;
 import br.livraria.dominio.Categoria;
 import br.livraria.dominio.Livro;
 
-
 public interface PesquisaDao {
-	
+
 	public List<Livro> pesquisarPorAutor(Long codigoAutor) throws SQLException;
+
 	public List<Livro> pesquisarPorTitulo(String titulo) throws SQLException;
+
 	public List<Livro> pesquisarPorCategoria(Long codigoCategoria) throws SQLException;
-	public List<Livro> pesquisarPorEditora(Integer codigoEditora) throws SQLException;
-	
+
+	public List<Livro> pesquisarPorEditora(Long codigoEditora) throws SQLException;
+
 }
